Split up concept files to allow for finer reuse
