+++ include/eve/concept/logical_simd.hpp
@@ -1,0 +1,31 @@
+//==================================================================================================
+/*
+  EVE - Expressive Vector Engine
+  Copyright : EVE Contributors & Maintainers
+  SPDX-License-Identifier: MIT
+*/
+//==================================================================================================
+#pragma once
+
+#include <eve/concept/simd.hpp>
+#include <eve/detail/meta.hpp>
+
+namespace eve
+{
+  template<typename Type> struct logical;
+
+  //================================================================================================
+  //! @concept logical_simd_value
+  //! @brief Specify that a type represents a logical SIMD value
+  //!
+  //! The concept `logical_simd_value<T>` is satisfied if and only if T is an instance of
+  //! eve::logical and its contents satisfy eve::plain_simd_value
+  //!
+  //! ## Example Types
+  //! - eve::logical<eve::wide<float>>
+  //! - eve::logical<eve::wide<int,eve::fixed<2>>>
+  //================================================================================================
+  template<typename T>
+  concept logical_simd_value  =   detail::instance_of<T,logical>
+                              &&  plain_simd_value<typename T::mask_type>;
+}
